+++ browser-extension/src/infrastructure/prun-api/data/material-categories.ts
@@ -1,5 +1,6 @@
 import { createEntityStore } from '@src/infrastructure/prun-api/data/create-entity-store';
 import { messages } from '@src/infrastructure/prun-api/data/api-messages';
+import { createMapGetter } from '@src/infrastructure/prun-api/data/create-map-getter';
 
 const store = createEntityStore<PrunApi.MaterialCategory>();
 const state = store.state;
@@ -11,6 +12,12 @@ messages({
   },
 });
 
+export const toSerializableCategoryName = (name: string) =>
+  name.replaceAll('(', '').replaceAll(')', '').toUpperCase();
+
+const getBySerializableName = createMapGetter(state.all, x => x.name, toSerializableCategoryName);
+
 export const materialCategoriesStore = {
   ...state,
+  getBySerializableName,
 };