+++ browser-extension/src/refined-prun.ts
@@ -66,7 +66,7 @@ import '@src/features/XIT/HEALTH.vue';
 import '@src/features/XIT/HELP';
 import '@src/features/XIT/HQUC.vue';
 import '@src/features/XIT/LISTS';
-import '@src/features/XIT/MAT.vue';
+import '@src/features/XIT/MATS.vue';
 import '@src/features/XIT/NOTE/NOTE.vue';
 import '@src/features/XIT/REP/REP.vue';
 import '@src/features/XIT/SET/SET.vue';