+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - `XIT FINBS`: Balance Statement
 - `XIT HQUC`: HQ Upgrade Calculator
 - `XIT CONTC`: Pending Contract Conditions
+- `XIT MATS`: List all materials (with filtering via parameters)
 - `XIT WEB`: Open any web page (Pro tip! Try `XIT WEB https://www.youtube.com/embed/dQw4w9WgXcQ`)
 - `XIT GIF`: Random GIF (The main reason for this one is `XIT GIF cats`)
 - `XIT FIN` is divided into multiple commands: